requirements.txt
Flask
Flask-Cors
python-dotenv
pdfplumber
python-docx
spacy
scikit-learn
numpy
pandas